Add support for 64-bit

+++ ZXingObjC/ZXDecodeHints.m
@@ -66,7 +66,7 @@
 }
 
 - (int)numberOfPossibleFormats {
-  return self.barcodeFormats.count;
+  return (int)self.barcodeFormats.count;
 }
 
 - (void)removePossibleFormat:(ZXBarcodeFormat)format {
